+++ src/lib.rs
@@ -17,24 +17,54 @@ mod options;
 pub mod win_service;
 
 pub use cli::Cli;
-pub use options::{Options, OptionsError};
+pub use options::{Format, Options, OptionsError};
 
-/// Wrapper around a [`CliResult`] that provides [`Termination`] support
-pub struct MainResult(CliResult);
+/// Wrapper around a [`CliResult`] that provides [`Termination`] support and [`Format`]ing.
+pub struct MainResult {
+    inner: CliResult,
+    format: Format,
+}
 
 impl MainResult {
-    pub const OK: MainResult = MainResult(Ok(()));
+    pub const OK: MainResult = MainResult {
+        inner: Ok(()),
+        format: Format::Shell,
+    };
+
+    /// Creates a new result that performs general shell formatting.
+    pub fn new(inner: CliResult) -> Self {
+        Self {
+            inner,
+            format: Format::Shell,
+        }
+    }
+
+    /// Converts to shell formatting for errors.
+    pub fn shell(self) -> Self {
+        Self {
+            inner: self.inner,
+            format: Format::Shell,
+        }
+    }
+
+    /// Converts to a JSON formatting for errors.
+    pub fn json(self) -> Self {
+        Self {
+            inner: self.inner,
+            format: Format::Json,
+        }
+    }
 }
 
 impl From<CliResult> for MainResult {
     fn from(res: CliResult) -> Self {
-        Self(res)
+        Self::new(res)
     }
 }
 
 impl From<OptionsError> for MainResult {
     fn from(x: OptionsError) -> Self {
-        Self(match x {
+        Self::new(match x {
             OptionsError::Config(x) => Err(CliError::Error(x)),
             OptionsError::Options(x) => match x.kind() {
                 // --help and --version should not actually exit with an error and instead display
@@ -57,13 +87,13 @@ impl From<OptionsError> for MainResult {
 
 impl From<anyhow::Error> for MainResult {
     fn from(x: anyhow::Error) -> Self {
-        Self(Err(CliError::Error(x)))
+        Self::new(Err(CliError::Error(x)))
     }
 }
 
 impl From<anyhow::Result<()>> for MainResult {
     fn from(res: anyhow::Result<()>) -> Self {
-        Self(res.map_err(CliError::Error))
+        Self::new(res.map_err(CliError::Error))
     }
 }
 
@@ -86,12 +116,22 @@ impl CliError {
 
 impl Termination for MainResult {
     fn report(self) -> ExitCode {
-        match self.0 {
+        match self.inner {
             Ok(_) => ExitCode::SUCCESS,
             Err(x) => match x {
                 CliError::Exit(code) => ExitCode::from(code),
                 CliError::Error(x) => {
-                    eprintln!("{x:?}");
+                    match self.format {
+                        Format::Shell => eprintln!("{x}"),
+                        Format::Json => println!(
+                            "{}",
+                            serde_json::to_string(&serde_json::json!({
+                                "type": "error",
+                                "msg": x.to_string(),
+                            }),)
+                            .expect("Failed to format error to JSON")
+                        ),
+                    }
                     ::log::error!("{x:?}");
                     ::log::logger().flush();
                     ExitCode::FAILURE